requirements.txt
langchain>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langgraph>=1.0.0
langchain-mcp-adapters>=0.1.14
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
httpx
beautifulsoup4
fastapi
uvicorn
sse-starlette
ddgs
